backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
motor>=2.5.1
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=0.19.0
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=3.2.0
email-validator>=1.1.3